requirements.txt
fastapi[all]
pillow
torch
torchvision
plotly
pandas